+++ data/updateMajors.js
@@ -7,19 +7,18 @@
 import * as db from './db.js';
 import { allMajors } from './majors.js';
 import Majors from '../model/Major.js';
-import e from 'cors';
 
 // addAllMajors();
 updateAllMajors();
 
-// Add ALL majors in allMajors array to production DB
+// add majors in allMajors array if not already in production DB
 async function addAllMajors() {
-  await db.connect(); // comment out this line before running jest test
+  await db.connect(); 
   let new_count = 0;
   for (let major of allMajors) {
-    const existing_major = await majors.findById(major);
+    const existing_major = await Majors.findById(major);
     if (!existing_major) {
-      majors
+      Majors
         .insert(major)
         .then((res) => {
           console.log(`added new degree: ${res.data}`);
@@ -30,16 +29,12 @@ async function addAllMajors() {
       console.log(`${major._id} already exists!`);
     }
   }
-  console.log(
-    console.log(
-      `Done adding ${new_count} new majors! Check DB to confirm the documents have been added to the collection`,
-    ),
-  );
+  console.log(`Done adding ${new_count} new majors! Check DB to confirm the documents have been added to the collection`);
 }
 
-// Add ALL majors in allMajors array to production DB
+// add new majors and update existing ones in production DB
 async function updateAllMajors() {
-  await db.connect(); // comment out this line before running jest test
+  await db.connect(); 
   let new_count = 0;
   let update_count = 0;
   for (let major of allMajors) {
@@ -58,19 +53,17 @@ async function updateAllMajors() {
       console.log(err);
     }
   }
-  console.log(
-    console.log(
-      `Done adding ${new_count} new majors and updating ${update_count} existing majors! Check DB to confirm.`,
-    ),
-  );
+  console.log(`Done adding ${new_count} new majors and updating ${update_count} existing majors! Check DB to confirm.`);
 }
 
 // Add a specified major in allMajors array to production DB
+// majorName is _id of the major
 async function addOneMajor(majorName) {
-  await db.connect(); // comment out this line before running jest test
+  await db.connect(); 
   let documentAdded = false;
   for (let major of allMajors) {
     if (major._id === majorName) {
+      // found specified major 
       await Majors.create(major)
         .then((majorDocument) => {
           documentAdded = true;
@@ -80,14 +73,10 @@ async function addOneMajor(majorName) {
           );
         })
         .catch((err) => console.log(err));
+      break; 
     }
   }
   if (!documentAdded) {
     console.log('No documents were added to the collection.');
   }
 }
-
-// module.exports = {
-//   addAllMajors,
-//   addOneMajor,
-// };
